chore(test): setup test for database class
-ensure all tests are passed

[Fixes #164350357]
