+++ .eslintrc
@@ -1,3 +1,9 @@
 {
-    "extends": "airbnb-base"
+    "extends": "airbnb-base",
+    "plugins": [
+        "mocha"
+    ],
+    "rules": {
+        "mocha/no-exclusive-tests": "error"
+    }
 }
