+++ package.json
@@ -4,7 +4,7 @@
   "description": "A wep app to share messages over the internet",
   "main": "server/app.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "mocha --require @babel/register test/*.js",
     "start": "nodemon --exec babel-node server/app.js"
   },
   "repository": {
@@ -21,11 +21,12 @@
     "@babel/core": "^7.3.4",
     "@babel/node": "^7.2.2",
     "@babel/preset-env": "^7.3.4",
-    "babel-core": "^6.26.3",
+    "@babel/register": "^7.0.0",
     "eslint": "^5.15.0",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-config-airbnb-base": "^13.1.0",
     "eslint-plugin-import": "^2.16.0",
+    "eslint-plugin-mocha": "^5.3.0",
     "nodemon": "^1.18.10"
   },
   "dependencies": {
